requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Vector Database and Search
qdrant-client==1.6.9
sentence-transformers==2.2.2

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
docx2txt==0.8

# Background Tasks
celery==5.3.4
redis==5.0.1

# Object Storage
minio==7.2.0

# ML and NLP
transformers==4.35.2
torch==2.1.1
spacy==3.7.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0